Migrate to TypeScript
