+++ apps/meteor/app/user-data-download/server/createDir.ts
@@ -1,0 +1,7 @@
+import { existsSync, mkdirSync } from 'fs';
+
+export const createDir = (folderName: string): void => {
+	if (!existsSync(folderName)) {
+		mkdirSync(folderName, { recursive: true });
+	}
+};
