--- apps/meteor/app/user-data-download/server/index.js
@@ -1,2 +1,0 @@
-import './cronProcessDownloads';
-import './exportDownload';
