+++ apps/meteor/app/user-data-download/server/makeZipFile.ts
@@ -1,0 +1,19 @@
+import { createWriteStream } from 'fs';
+
+import archiver from 'archiver';
+
+export const makeZipFile = (folderToZip: string, targetFile: string): Promise<void> => {
+	return new Promise((resolve, reject) => {
+		const output = createWriteStream(targetFile);
+
+		const archive = archiver('zip');
+
+		output.on('close', () => resolve());
+
+		archive.on('error', (err) => reject(err));
+
+		archive.pipe(output);
+		archive.directory(folderToZip, false);
+		archive.finalize();
+	});
+};
