+++ apps/meteor/app/user-data-download/server/startFile.ts
@@ -1,0 +1,5 @@
+import { writeFileSync } from 'fs';
+
+export const startFile = (fileName: string, content: string | NodeJS.ArrayBufferView) => {
+	writeFileSync(fileName, content);
+};
