+++ apps/meteor/app/user-data-download/server/writeToFile.ts
@@ -1,0 +1,5 @@
+import { appendFileSync } from 'fs';
+
+export const writeToFile = (fileName: string, content: string | Uint8Array) => {
+	appendFileSync(fileName, content);
+};
