+++ apps/meteor/server/importPackages.ts
@@ -84,7 +84,6 @@ import '../app/threads/server';
 import '../app/tokenpass/server';
 import '../app/ui-master/server';
 import '../app/ui-vrecord/server';
-import '../app/user-data-download/server';
 import '../app/webdav/server';
 import '../app/webrtc/server';
 import '../app/wordpress/server';
