+++ apps/meteor/server/startup/cron.js
@@ -7,6 +7,7 @@ import { statsCron } from '../cron/statistics';
 import { npsCron } from '../cron/nps';
 import { federationCron } from '../cron/federation';
 import { videoConferencesCron } from '../cron/videoConferences';
+import { userDataDownloadsCron } from '../cron/userDataDownloads';
 
 const logger = new Logger('SyncedCron');
 
@@ -23,6 +24,7 @@ Meteor.defer(function () {
 	npsCron(SyncedCron);
 	federationCron(SyncedCron);
 	videoConferencesCron(SyncedCron);
+	userDataDownloadsCron(SyncedCron);
 
 	SyncedCron.start();
 });
